add testing container and other simple tests

+++ mint/internal/lightning/utils.go
@@ -2,13 +2,13 @@ package lightning
 
 import "github.com/lightningnetwork/lnd/lnrpc"
 
-func GetAverageRouteFee(routes []*lnrpc.Route) int64 {
-	var fees int64
-	var amount_routes int
+func GetAverageRouteFee(routes []*lnrpc.Route) uint64 {
+	var fees uint64
+	var amount_routes uint64
 
 	for _, route := range routes {
-		fees += route.TotalFeesMsat
+		fees += uint64(route.TotalFeesMsat)
 		amount_routes += 1
 	}
-	return fees / int64(amount_routes)
+	return fees / amount_routes
 }
